functions/compile-gif/requirements.txt
Pillow
google-cloud-storage
apng
